django-demo/backend/requirements.txt
Django==3.2.14
graphene-django==2.15.0
